README polish, new embedding models (#25)
* get_info() method for QPM model

* documented number of parameters

* Suppressed tokenizer max-length warning

* Fixed issue with OpenAI requests in batches.

* OpenAI results

* Second README for Pypi.

* Support for loading embedding models as QPMMs.

* Reusing default embedding model across multiple calls to create().

* Default tokenizer also cached

* weak references for cached embedding model and tokenizer

* Added unit tests

* one more unit test

* one more unit test, signature correction

* New models

* More robust use of cached model files

* addressed OOM with big batch and p5 emb model

* Redid section on QPM evaluations to use best embedding model as QPM
model.

* Replaced name of undocumented QPM model

+++ goodai/ltm/mem/config.py
@@ -9,7 +9,7 @@ class TextMemoryConfig:
     The capacity of the memory's chunk queue, i.e. the maximum number of chunks it can hold.
     """
 
-    reranking_k_factor: int
+    reranking_k_factor: float
     """
     When a reranking mechanism is available, the value of k passed to the query function is
     multiplied by reranking_k_factor and the resulting number of chunks is retrieved from
@@ -19,4 +19,4 @@ class TextMemoryConfig:
     def __init__(self):
         self.chunk_capacity = 24  # Tokens
         self.queue_capacity = 5000  # Chunks
-        self.reranking_k_factor = 10
+        self.reranking_k_factor = 10.0
